Added example notebook for tracing and deploying CLIP model to opensearch (#283)
* Example of deploying CLIPTextModel to OpenSearch

This notebook provides a walkthrough for users to trace, register, and deploy a CLIPTextModel from a local file. CLIPTextModel can be used with the neural search plugin to generate embeddings of documents and ingest time and of user queries at search time. 


Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>

* Update index.rst

Added line for new example notebook demo_deploy_cliptextmodel

Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>

* Delete docs/source/examples/demo_deploy_cliptextmodel.ipynb

Deleted notebook, will upload new version

Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>

* Uploaded Demo Notebook 

This notebook provides a walkthrough for users to trace, register, and deploy a CLIPTextModel from a local file. CLIPTextModel can be used with the Neural Search plugin to generate embeddings of documents and ingest time and of user queries at search time. 

Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>

* Update CHANGELOG.md

Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>

* Update CHANGELOG.md

Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>

---------

Signed-off-by: patrickbarnhart <patrick@dahlialabs.ai>