Update rag_setup.py to remove neural search line and serverless
Removed references to serverless. Now the code only supports managed and open-source service types.
Removed the menu option for serverless.
For search methods, removed “neural” and only use “semantic” search.
Added a prompt to choose between semantic search with LLM and semantic search without LLM.
If semantic with LLM is chosen and the service type is managed, the code will prompt for LLM configuration.
If open-source is chosen, we skip AWS/Bedrock configurations and do not prompt for LLM registration since that requires AWS Bedrock.

Signed-off-by: hmumtazz <144855436+hmumtazz@users.noreply.github.com>